update images column in workload detail pod table
