+++ components/formatter/PodImages.vue
@@ -1,19 +1,46 @@
 <script>
+import { mapGetters } from 'vuex';
+import { WORKLOAD_TYPES } from '@/config/types';
 export default {
   props: {
     value: {
-      type:     Array,
+      type:     [Array, String],
       default: () => {
         return [];
       }
+    },
+
+    row: {
+      type:    Object,
+      default: () => {
+        return {};
+      }
     }
   },
-  computed: {
-    images() {
-      return this.value.map(container => container.image)
-        .filter(image => !image.startsWith('sha256'));
+
+  async fetch() {
+    if (Object.values(WORKLOAD_TYPES).includes(this.row.type)) {
+      const pods = await this.row.pods();
+
+      this.images = pods.reduce((images, pod) => {
+        const podImages = pod.spec.containers.reduce((all, container) => {
+          all.push(container.image);
+
+          return all;
+        }, []);
+
+        images.push(...podImages);
+
+        return images;
+      }, []);
     }
-  }
+  },
+
+  data() {
+    return { images: this.value };
+  },
+
+  computed: { ...mapGetters({ t: 'i18n/t' }) }
 
 };
 </script>
@@ -21,6 +48,6 @@ export default {
 <template>
   <span>
     <span>{{ images[0] }}</span><br>
-    <span v-if="images.length-1>0" class="plus-more">+{{ images.length-1 }} more</span>
+    <span v-if="images.length-1>0" class="plus-more">{{ t('generic.plusMore', {n:images.length-2}) }}</span>
   </span>
 </template>
